fix(emails): restore '11 dias' in Day 3 engagement + split trial.py to pass LOC gate
Backend Tests failure: test_engagement_email_says_11_dias asserted '11 dias'
in render_trial_engagement_email output. Belgray rewrite (#1005) dropped the
days_remaining mention. Add days_remaining: int = 11 param (Day 3 → 14-3=11)
and weave naturally into the close paragraph: "Ainda faltam {N} dias do seu
trial, dá pra explorar com calma." Preserves Belgray friend-text voice while
restoring the STORY-321 contract verified by 4 days-remaining assertions.

LOC gate failure: trial.py grew 912 → 1123 LOC (+23%, over the +20% growth
warning threshold) and the workflow's GITHUB_OUTPUT step blew up on the
WARNINGS_JSON multiline value, escalating warning to step failure. Extract
helpers (_founders_*, _founder_signature, _format_brl, _stats_block,
_unsubscribe_block, _preheader + 4 pricing constants) into new
_trial_helpers.py and re-export from trial.py for backwards-compatible
imports. trial.py is now 970 LOC (+6% growth, well under +20%).

Verified:
- pytest tests/test_trial_14_days.py tests/test_trial_email_sequence.py
  tests/test_trial_emails.py — 160/160 passed
- pytest -k "trial or email" — 745 passed, 0 failed (5min full sweep)
- ruff check on _trial_helpers.py — clean
